lief::coff

Module symbol

Structs

AuxiliaryBfAndEf

AuxiliaryCLRToken

AuxiliaryFile

This auxiliary symbol represents a filename (auxiliary format 4)

AuxiliaryFunctionDefinition

This auxiliary symbols marks the beginning of a function definition.

AuxiliarySectionDefinition

This auxiliary symbol exposes information about the associated section.

AuxiliarySymbol

AuxiliaryWeakExternal

“Weak externals” are a mechanism for object files that allows flexibility at link time. A module can contain an unresolved external symbol (sym1), but it can also include an auxiliary record that indicates that if sym1 is not present at link time, another external symbol (sym2) is used to resolve references instead.

Symbol

Structure that represents a COFF symbol.

Enums

AuxiliarySymbols

Class that represents an auxiliary symbols.

BaseType

Characteristics

ComdatSelection

ComplexType

StorageClass

Type Aliases

ItAuxiliarySymbols