lief::dsc

Struct Dylib

pub struct Dylib<'a> { /* private fields */ }

This structure represents a library embedded in a dyld shared cache. It mirrors the original dyld_cache_image_info structure.

Implementations

impl Dylib<'_>

pub fn path(&self) -> String

Original path of the library (e.g. /usr/lib/libcryptex.dylib)

pub fn address(&self) -> u64

In-memory address of the library

pub fn modtime(&self) -> u64

Modification time of the library matching stat.st_mtime, or 0

pub fn inode(&self) -> u64

File serial number matching stat.st_ino or 0

Note that for shared cache targeting iOS, this value can hold a hash of the path (if modtime is set to 0)

pub fn padding(&self) -> u64

Padding alignment value (should be 0)

pub fn get(&self) -> Option<Binary>

Get a Binary representation for this Dylib.

One can use this function to write back the Mach-O binary on the disk:

dylib.get().expect("Can't extract").write("liblockdown.dylib");

pub fn get_with_opt(&self, opt: &ExtractOpt) -> Option<Binary>

Get a Binary representation for this Dylib with the provided ExtractOpt options.