lief::elf

Module dynamic

Structs

Array

Structure that represents a dynamic entry associated with an array (e.g. DT_INIT_ARRAY)

DtFlags

Flags

Structure that represents a dynamic flag entry: DT_FLAGS or DT_FLAGS_1

Generic

Generic structure for the dynamic entries whose DynamicEntry::value can be interpreted as is.

Library

Structure that represents a dynamic entry associated with a library name (e.g. DT_NEEDED)

Rpath

Structure that represents a dynamic entry associated with the rpath info

RunPath

Structure that represents a dynamic entry associated with the runpath info

SharedObject

Structure that represents a dynamic entry associated with the name of a library (DT_SONAME)

Enums

Entries

Enum that represents the different variants of a dynamic entry

Tag

Traits

DynamicEntry

Trait shared by all the Entries

Type Aliases

DynamicEntries