Enum lief::elf::section::Type

pub enum Type {

    SHT_NULL,
    PROGBITS,
    SYMTAB,
    STRTAB,
    RELA,
    HASH,
    DYNAMIC,
    NOTE,
    NOBITS,
    REL,
    SHLIB,
    DYNSYM,
    INIT_ARRAY,
    FINI_ARRAY,
    PREINIT_ARRAY,
    GROUP,
    SYMTAB_SHNDX,
    RELR,
    ANDROID_REL,
    ANDROID_RELA,
    LLVM_ADDRSIG,
    ANDROID_RELR,
    GNU_ATTRIBUTES,
    GNU_HASH,
    GNU_VERDEF,
    GNU_VERNEED,
    GNU_VERSYM,
    ARM_EXIDX,
    ARM_PREEMPTMAP,
    ARM_ATTRIBUTES,
    ARM_DEBUGOVERLAY,
    ARM_OVERLAYSECTION,
    HEX_ORDERED,
    X86_64_UNWIND,
    MIPS_REGINFO,
    MIPS_OPTIONS,
    MIPS_ABIFLAGS,
    UNKNOWN(u64),

}

Variants

SHT_NULL

No associated section (inactive entry)

PROGBITS

Program-defined contents.

SYMTAB

Symbol table

STRTAB

String table

RELA

Relocation entries; explicit addends.

HASH

Symbol hash table.

DYNAMIC

Information for dynamic linking.

NOTE

Information about the file.

NOBITS

Data occupies no space in the file.

REL

Relocation entries; no explicit addends.

SHLIB

Reserved

DYNSYM

Symbol table.

INIT_ARRAY

Pointers to initialization functions.

FINI_ARRAY

Pointers to termination functions.

PREINIT_ARRAY

Pointers to pre-init functions.

GROUP

Section group.

SYMTAB_SHNDX

Indices for SHN_XINDEX entries.

RELR

Relocation entries; only offsets.

ANDROID_REL

Packed relocations (Android specific).

ANDROID_RELA

Packed relocations (Android specific).

LLVM_ADDRSIG

This section is used to mark symbols as address-significant.

ANDROID_RELR

New relr relocations (Android specific).

GNU_ATTRIBUTES

Object attributes.

GNU_HASH

GNU-style hash table.

GNU_VERDEF

GNU version definitions.

GNU_VERNEED

GNU version references.

GNU_VERSYM

GNU symbol versions table.

ARM_EXIDX

Exception Index table

ARM_PREEMPTMAP

BPABI DLL dynamic linking pre-emption map

ARM_ATTRIBUTES

Object file compatibility attributes

ARM_DEBUGOVERLAY

ARM_OVERLAYSECTION

HEX_ORDERED

Link editor is to sort the entries in this section based on their sizes

X86_64_UNWIND

Unwind information

MIPS_REGINFO

Register usage information

MIPS_OPTIONS

General options

MIPS_ABIFLAGS

ABI information

UNKNOWN(u64)

Implementations

impl Type

pub fn from_value(value: u64) -> Self