lief::elf

Struct Symbol

pub struct Symbol<'a> { /* private fields */ }

Structure which reprents an ELF symbol

Implementations

impl Symbol<'_>

pub fn get_type(&self) -> Type

The symbol’s type provides a general classification for the associated entity

pub fn binding(&self) -> Binding

The symbol’s binding determines the linkage visibility and behavior

pub fn information(&self) -> u8

This member specifies the symbol’s type and binding attributes.

pub fn other(&self) -> u8

Alias for Symbol::visibility

pub fn section_idx(&self) -> u16

ELF Section index associated with the symbol

pub fn visibility(&self) -> Visibility

Symbol visibility

pub fn section(&self) -> Option<Section<'_>>

Section associated with the symbol (if any)

pub fn symbol_version(&self) -> Option<SymbolVersion<'_>>

Return the SymbolVersion associated with this symbol (if any)

pub fn demangled_name(&self) -> String

Try to demangle the symbol or return an empty string if it is not possible