lief::generic

Trait DebugInfo

pub trait DebugInfo {
    // Provided method
    fn find_function_address(&self, name: &str) -> Option<u64> { ... }
}

This class provides a generic interface for accessing debug information from different formats such as DWARF and PDB.

Users can use this interface to access high-level debug features like resolving function addresses.

See: crate::pdb::DebugInfo, crate::dwarf::DebugInfo

Provided Methods

fn find_function_address(&self, name: &str) -> Option<u64>

Attempt to resolve the address of the function specified by name.