lief::generic

Trait Symbol

pub trait Symbol {
    // Provided methods
    fn name(&self) -> String { ... }
    fn value(&self) -> u64 { ... }
    fn size(&self) -> u64 { ... }
    fn set_name(&mut self, name: &str) { ... }
    fn set_value(&mut self, value: u64) { ... }
    fn set_size(&mut self, size: u64) { ... }
}

Trait shared by all the symbols in executable formats

Provided Methods

fn name(&self) -> String

Symbol’s name

fn value(&self) -> u64

Symbol’s value whose interpretation depends on the symbol’s kind. Usually this is the address of the symbol though.

fn size(&self) -> u64

Size of the symbol (can be 0)

fn set_name(&mut self, name: &str)

fn set_value(&mut self, value: u64)

fn set_size(&mut self, size: u64)