lief::macho::commands

Struct DyldChainedFixups

pub struct DyldChainedFixups<'a> { /* private fields */ }

Structure that represents the LC_DYLD_CHAINED_FIXUPS command

This command aims at providing rebase and binding information like the DyldInfo’s bytecode. Compared to the DyldInfo bytecode, these chained fixups are taking less space.

Implementations

impl DyldChainedFixups<'_>

pub fn data_offset(&self) -> u32

Offset of the LC_DYLD_CHAINED_FIXUPS chained payload. This offset should point in the __LINKEDIT segment

pub fn data_size(&self) -> u32

Size of the LC_DYLD_CHAINED_FIXUPS payload.

pub fn fixups_version(&self) -> u32

Chained fixups version. The loader (dyld v852.2) checks that this value is set to 0

pub fn starts_offset(&self) -> u32

offset of dyld_chained_starts_in_image in chain_data

pub fn imports_offset(&self) -> u32

Offset of imports table in chain data

pub fn symbols_offset(&self) -> u32

Offset of symbol strings in chain data

pub fn imports_count(&self) -> u32

Number of imported symbol names

pub fn symbols_format(&self) -> u32

The compression algorithm (if any) used to store the symbols 0 means uncompressed while 1 means zlib compressed.

As far of the version v852.2 of dyld loader, it only supports uncompressed format

pub fn imports_format(&self) -> CHAINED_FORMAT

The format of the imports

pub fn bindings(&self) -> Bindings<'_>

Iterator over the bindings crate::macho::BindingInfo::Chained associated with this command

pub fn payload(&self) -> &[u8]

Return the raw content of the command