Trait lief::macho::relocation::RelocationBase

pub trait RelocationBase {
    // Provided methods
    fn is_pc_relative(&self) -> bool { ... }
    fn symbol(&self) -> Option<Symbol<'_>> { ... }
    fn section(&self) -> Option<Section<'_>> { ... }
    fn segment(&self) -> Option<Segment<'_>> { ... }
}

Trait shared by all the relocations defined in Relocation

Provided Methods

fn is_pc_relative(&self) -> bool

Indicates whether the item containing the address to be relocated is part of a CPU instruction that uses PC-relative addressing.

For addresses contained in PC-relative instructions, the CPU adds the address of the instruction to the address contained in the instruction.

fn symbol(&self) -> Option<Symbol<'_>>

Symbol associated with the relocation (if any)

fn section(&self) -> Option<Section<'_>>

Section associated with the section (if any)

fn segment(&self) -> Option<Segment<'_>>

Segment command associated with the relocation (if any)