lief::macho::section

Enum Type

pub enum Type {

    REGULAR,
    ZEROFILL,
    CSTRING_LITERALS,
    S_4BYTE_LITERALS,
    S_8BYTE_LITERALS,
    LITERAL_POINTERS,
    NON_LAZY_SYMBOL_POINTERS,
    LAZY_SYMBOL_POINTERS,
    SYMBOL_STUBS,
    MOD_INIT_FUNC_POINTERS,
    MOD_TERM_FUNC_POINTERS,
    COALESCED,
    GB_ZEROFILL,
    INTERPOSING,
    S_16BYTE_LITERALS,
    DTRACE_DOF,
    LAZY_DYLIB_SYMBOL_POINTERS,
    THREAD_LOCAL_REGULAR,
    THREAD_LOCAL_ZEROFILL,
    THREAD_LOCAL_VARIABLES,
    THREAD_LOCAL_VARIABLE_POINTERS,
    THREAD_LOCAL_INIT_FUNCTION_POINTERS,
    INIT_FUNC_OFFSETS,
    UNKNOWN(u64),

}

Variants

REGULAR

Regular section.

ZEROFILL

Zero fill on demand section.

CSTRING_LITERALS

Section with literal C strings.

S_4BYTE_LITERALS

Section with 4 byte literals.

S_8BYTE_LITERALS

Section with 8 byte literals.

LITERAL_POINTERS

Section with pointers to literals.

NON_LAZY_SYMBOL_POINTERS

Section with non-lazy symbol pointers.

LAZY_SYMBOL_POINTERS

Section with lazy symbol pointers.

SYMBOL_STUBS

Section with symbol stubs, byte size of stub in the Reserved2 field.

MOD_INIT_FUNC_POINTERS

Section with only function pointers for initialization.

MOD_TERM_FUNC_POINTERS

Section with only function pointers for termination.

COALESCED

Section contains symbols that are to be coalesced.

GB_ZEROFILL

Zero fill on demand section (that can be larger than 4 gigabytes).

INTERPOSING

Section with only pairs of function pointers for interposing.

S_16BYTE_LITERALS

Section with only 16 byte literals.

DTRACE_DOF

Section contains DTrace Object Format.

LAZY_DYLIB_SYMBOL_POINTERS

Section with lazy symbol pointers to lazy loaded dylibs.

THREAD_LOCAL_REGULAR

Thread local data section.

THREAD_LOCAL_ZEROFILL

Thread local zerofill section.

THREAD_LOCAL_VARIABLES

Section with thread local variable structure data.

THREAD_LOCAL_VARIABLE_POINTERS

Section with pointers to thread local structures.

THREAD_LOCAL_INIT_FUNCTION_POINTERS

Section with thread local variable initialization pointers to functions.

INIT_FUNC_OFFSETS

Section with 32-bit offsets to initializer functions

UNKNOWN(u64)