lief

Module objc

Module for Objective-C metadata

This module contains the support of Objective-C metadata embedded in Mach-O crate::macho::Binary.

One can access this metadata using the crate::macho::Binary::objc_metadata function

fn inspect_objc(macho: &lief::macho::Binary) {
    let metadata: lief::objc::Metadata = macho.objc_metadata().expect("Can't access metadata");
    for class in metadata.classes() {
        println!("{}", class.to_decl());
    }
}

Modules

class

decl_opt

ivar

metadata

method

property

protocol

Structs

Class

This class represents an Objective-C class (@interface)

DeclOpt

This structure wraps options to tweak the generated output of functions like crate::objc::Metadata::to_decl_with_opt

IVar

This structure represents an instance variable (ivar)

Metadata

This structure is the main interface to inspect Objective-C metadata

Method

This structure represents an Objective-C Method

Property

This class represents a @property in Objective-C

Protocol

This class represents an Objective-C @protocol