Module lief::pe

Module for the PE format

Modules

binary

code_integrity

data_directory

PE Data Directory module

debug

Module that wraps the different debug information structure we can find in a PE binary.

delay_import

PE Delayed import module

export

PE export module

headers

This module exposes the different headers that we can find in a PE binary. It includes:

import

This module represents PE’s Imports

load_configuration

relocation

This module includes the different structures related to the relocation process in a PE binary

resources

This module contains the different structures involved in the PE’s resource tree

rich_header

PE Rich Header module

section

PE section module

signature

This module wraps the PKCS #7 PE authenticode signature

tls

This module represents the PE’s Thread Local Storage (TLS)

Structs

Binary

This is the main interface to read and write PE binary attributes.

DataDirectory

DelayImport

DosHeader

Structure which represents the DosHeader, the first structure presents at the beginning of a PE file.

Export

Header

Structure that represents the PE header (which follows the DosHeader)

Import

OptionalHeader

Structure which represents the PE OptionalHeader (after Header).

Relocation

Class which represents the Base Relocation Block We usually find this structure in the .reloc section

ResourcesManager

RichEntry

RichHeader

Structure which represents the not-so-documented rich header

Section

This structure defines a regular PE section.

Signature

TLS

Enums

Algorithms

ResourceNode