lief::pe

Module debug

Module that wraps the different debug information structure we can find in a PE binary.

Structs

CodeView

Structure that represents the (generic) Debug CodeView (IMAGE_DEBUG_TYPE_CODEVIEW).

CodeViewPDB

CodeView PDB specialization

ExDllCharacteristics

This structure represents the IMAGE_DEBUG_TYPE_EX_DLLCHARACTERISTICS debug entry

ExtendedDLLCharacteristics

FPO

This structure represents the IMAGE_DEBUG_TYPE_FPO debug entry

FPOEntry

This structure represents the stack frame layout for a x86 function when frame pointer omission (FPO) optimization is used.

Generic

PDBChecksum

This structure represents the PDB Checksum debug entry which is essentially an array of bytes representing the checksum of the PDB content.

Pogo

This structure represents a Profile Guided Optimization entry from the debug directory (IMAGE_DEBUG_TYPE_POGO).

PogoEntry

Structure which reprents an entry in the Pogo debug structure

Repro

This structure represents a reproducible build entry from the debug directory (IMAGE_DEBUG_TYPE_REPRO)

VCFeature

This structure represents the IMAGE_DEBUG_TYPE_VC_FEATURE debug entry

Enums

ChecksumAlgorithm

Entries

This enum exposes the different debug entries that can be found in the debug DataDirectory.

Type

Traits

DebugEntry

Type Aliases

FPOEntries

PogoEntries