Trait lief::pe::debug::DebugEntry

pub trait DebugEntry {
    // Provided methods
    fn characteristics(&self) -> u32 { ... }
    fn timestamp(&self) -> u32 { ... }
    fn major_version(&self) -> u16 { ... }
    fn minor_version(&self) -> u16 { ... }
    fn get_type(&self) -> Type { ... }
    fn sizeof_data(&self) -> u32 { ... }
    fn addressof_rawdata(&self) -> u32 { ... }
    fn pointerto_rawdata(&self) -> u32 { ... }
}

Provided Methods

fn characteristics(&self) -> u32

Reserved should be 0

fn timestamp(&self) -> u32

The time and date when the debug data was created.

fn major_version(&self) -> u16

The major version number of the debug data format.

fn minor_version(&self) -> u16

The minor version number of the debug data format.

fn get_type(&self) -> Type

The format of the debugging information

fn sizeof_data(&self) -> u32

Size of the debug data

fn addressof_rawdata(&self) -> u32

Address of the debug data relative to the image base

fn pointerto_rawdata(&self) -> u32

File offset of the debug data