lief::pe

Module exception_x64

Module related to the x86_64 exception unwinding support

Structs

Alloc

This structure represents a stack-allocation operation (UnwindOpcodes::ALLOC_SMALL or UnwindOpcodes::ALLOC_LARGE).

Epilog

Describes the function’s epilog

PushMachFrame

Push a machine frame

PushNonVol

Push a nonvolatile integer register, decrementing RSP by 8

RuntimeFunction

This structure represents an entry in the exception table (.pdata section) for the x86-64 architecture.

SaveNonVolatile

Save a nonvolatile integer register on the stack using a MOV instead of a PUSH.

SaveXMM128

SetFPReg

Establish the frame pointer register by setting the register to some offset of the current RSP

Spare

UnwindFlags

UnwindInfo

This structure represents the UNWIND_INFO which records the effects a function has on the stack pointer, and where the nonvolatile registers are saved on the stack.

Enums

Opcodes

The different x86_64 unwind opcodes.

UnwindOpcodes

UnwindReg

Traits

Opcode

Trait shared by all Opcodes

Type Aliases

OpcodesIterator