lief::pe

Module resources

This module contains the different structures involved in the PE’s resource tree

One can access this tree using either: crate::pe::Binary::resources or by parsing raw bytes with Node::from_slice. The Manager provides a logical API over the resource tree to access specifics parts of the tree.

Structs

Accelerator

Represents a PE accelerator resource entry

Data

Directory

FixedFileInfo

Represents fixed file information from a version resource

Icon

Represents a PE icon resource

Manager

This manager abstracts the tree representation to provide a comprehensive API over the information wrapped by the resources tree.

ResourceVar

Represents a Var entry within VarFileInfo

StringEntry

Represents a string table entry from the ResourcesManager

StringFileInfo

Represents a StringFileInfo structure from a version resource

VarFileInfo

Represents a VarFileInfo structure from a version resource

Version

Represents a PE version resource

VersionStringTable

Represents a string table within a StringFileInfo

VersionStringTableEntry

Represents a string table entry (key/value) from a StringFileInfo

Enums

Node

This enum represents a node in the resource tree which can be either: a directory node or a data (leaf) node.

Types

Traits

NodeBase

Trait that is shared by both Node::Data and Node::Directory.

Type Aliases

Accelerators

Children

Icons

StringFileInfoChildren

StringTableEntries

VarFileInfoVars

VersionStringTableEntries

Versions