Module lief::pe::signature::attributes

Structs

ContentType

Interface over the structure described by the OID 1.2.840.113549.1.9.3 (PKCS #9)

GenericType

Interface over an attribute for which the internal structure is not supported by LIEF

MsCounterSign

This structure exposes the MS Counter Signature attribute

MsManifestBinaryID

Interface over the structure described by the OID 1.3.6.1.4.1.311.10.3.28 (szOID_PLATFORM_MANIFEST_BINARY_ID)

MsSpcNestedSignature

Interface over the structure described by the OID 1.3.6.1.4.1.311.2.4.1

MsSpcStatementType

Interface over the structure described by the OID 1.3.6.1.4.1.311.2.1.11

PKCS9AtSequenceNumber

Interface over the structure described by the OID 1.2.840.113549.1.9.25.4 (PKCS #9)

PKCS9CounterSignature

Interface over the structure described by the OID 1.2.840.113549.1.9.6 (PKCS #9)

PKCS9MessageDigest

Interface over the structure described by the OID 1.2.840.113549.1.9.4 (PKCS #9)

PKCS9SigningTime

Interface over the structure described by the OID 1.2.840.113549.1.9.5 (PKCS #9)

SigningCertificateV2

SpcRelaxedPeMarkerCheck

SpcSpOpusInfo

Interface over the structure described by the OID 1.3.6.1.4.1.311.2.1.12

Enums

Attribute

This enum exposes the different attributes that can be wrapped in a crate::pe::signature::SignerInfo

Type Aliases

MsCounterCertificates

MsCounterSigners