lief::pe::signature

Enum KeyUsage

pub enum KeyUsage {
    DIGITAL_SIGNATURE,
    NON_REPUDIATION,
    KEY_ENCIPHERMENT,
    DATA_ENCIPHERMENT,
    KEY_AGREEMENT,
    KEY_CERT_SIGN,
    CRL_SIGN,
    ENCIPHER_ONLY,
    DECIPHER_ONLY,
    UNKNOWN(u32),
}

Key usage as defined in RFC #5280 - section-4.2.1.3

Variants

DIGITAL_SIGNATURE

The key is used for digital signature

NON_REPUDIATION

The key is used for digital signature and to protect against falsely denying some action

KEY_ENCIPHERMENT

The key is used for enciphering private or secret keys

DATA_ENCIPHERMENT

The key is used for directly enciphering raw user data without an intermediate symmetric cipher

KEY_AGREEMENT

The key is used for key agreement (e.g. with Diffie-Hellman)

KEY_CERT_SIGN

The key is used for verifying signatures on public key certificates

CRL_SIGN

The key is used for verifying signatures on certificate revocation lists

ENCIPHER_ONLY

In association with KEY_AGREEMENT, the key is only used for enciphering data

DECIPHER_ONLY

In association with KEY_AGREEMENT, the key is only used for deciphering data

UNKNOWN(u32)