lief::pe

Struct DosHeader

pub struct DosHeader<'a> { /* private fields */ }

Structure which represents the DosHeader, the first structure presents at the beginning of a PE file.

Most of the attributes of this structures are no longer relevant

Implementations

impl<'a> DosHeader<'a>

pub fn magic(&self) -> u16

Magic bytes identifying a DOS/PE binary

pub fn used_bytes_in_last_page(&self) -> u16

pub fn file_size_in_pages(&self) -> u16

pub fn numberof_relocation(&self) -> u16

pub fn header_size_in_paragraphs(&self) -> u16

pub fn minimum_extra_paragraphs(&self) -> u16

pub fn maximum_extra_paragraphs(&self) -> u16

pub fn initial_relative_ss(&self) -> u16

pub fn initial_sp(&self) -> u16

pub fn checksum(&self) -> u16

pub fn initial_ip(&self) -> u16

pub fn initial_relative_cs(&self) -> u16

pub fn addressof_relocation_table(&self) -> u16

pub fn overlay_number(&self) -> u16

pub fn reserved(&self) -> Vec<u16>

pub fn oem_id(&self) -> u16

pub fn oem_info(&self) -> u16

pub fn reserved2(&self) -> Vec<u16>

pub fn addressof_new_exeheader(&self) -> u32

Return the offset to the Header structure.