lief

Module runtime

This module contains all LIEF’s runtime API

Modules

android

Android-specific API

assembler

disassembler

host

linux

Linux-specific API

memory

module

osx

process

windows

Structs

Host

This structure represents the current host.

Memory

Process

This structure represents the current process and provides functions to query process-level information.

Enums

Arch

Modules

Enum for the modules whose value depends on the current platform

Platform

Traits

Module

This trait is shared by all system-specific modules. It provides a common interface for accessing module metadata.

Functions

arch

Architecture for which the runtime is compiled

assemble

Assemble the provided assembly code at the specified (absolute) virtual address.

assemble_with_config

Same as assemble but this function takes an extra AssemblerConfig that is used to configure the assembly engine: dialect, symbols definitions.

disassemble

Start disassembling instructions at the given absolute virtual address.

enabled

Whether runtime features are enabled

module_from_addr

Find the module that encompasses the given virtual address (absolute)

module_from_name

Find the module with the given name

module_from_path

Find the module with the given name

modules

Return an iterator over the different modules loaded in the current process.

platform

Platform for which the runtime is compiled