lief::runtime::android

Struct Module

pub struct Module { /* private fields */ }

This structure represents an Android-specific module

Implementations

impl Module

pub fn from_handle(handle: *const c_void) -> Option<Module>

Instantiate a Module from the given dlopen handle.

pub fn handle(&self) -> *const c_void

Return the dlopen handle for this library as an opaque pointer.

Return a null pointer if the function fails or if the handler can’t be found.

pub fn dlsym(&self, name: String) -> *const c_void

Resolve the symbol with the given name for the current module

pub fn parse_from_path(&self) -> Option<Binary>

Parse the ELF module from its path on the filesystem

pub fn parse_from_path_with_config( &self, config: &ParserConfig, ) -> Option<Binary>

Parse the ELF module from its path on the filesystem and given the parser configuration

pub fn parse_from_memory(&self) -> Option<Binary>

Parse the ELF module from memory

pub fn parse_from_memory_with_config( &self, config: &ParserConfig, ) -> Option<Binary>

Parse the ELF module from memory with the given configuration