lief::runtime

Module windows

Modules

host

ldr_data_table_entry

module

peb

process

Structs

Host

This structure exposes Windows-specific host information.

LdrDataTableEntry

This structure exposes a user-friendly interface over a LDR_DATA_TABLE_ENTRY, the structure used by the Windows loader to describe a module loaded in the current process.

Module

This structure represents a Windows-specific module

PEB

This structure exposes a user-friendly interface over the Process Environment Block (PEB) of the current process.

Process

This structure exposes Windows-specific API for the current process.

Version

Represents a Windows version number

Functions

dlopen

Load the module (DLL) with the given name into the current process and wrap the resulting handle in a Module.

find_module

Try to get the Module with the given name.

Type Aliases

LdrDataTableEntries