Module lief::assembly

Assembly/Disassembly Module

Getting Started

One can start disassembling code by using the different functions exposes in the crate::generic::Binary trait:

fn disassemble(target: &dyn lief::generic::Binary) {
    for inst in target.disassemble_symbol("_entrypoint") {
        println!("{}", inst.to_string());
    }
}

An instruction is represented by the enum: Instructions which implements the trait Instruction.

For architecture-dependant API, you can check the following structures:

aarch64::Instruction

x86::Instruction

arm::Instruction

powerpc::Instruction

riscv::Instruction

mips::Instruction

ebpf::Instruction

Modules

aarch64

AArch64 architecture-related namespace

arm

ARM architecture-related namespace

ebpf

eBPF architecture-related namespace

instruction

Module related to assembly instructions

mips

Mips architecture-related namespace

powerpc

PowerPC architecture-related namespace

riscv

RISC-V architecture-related namespace

x86

x86/x86-64 architecture-related namespace

Enums

Instructions

All instruction variants supported by LIEF

Traits

Instruction

This trait is shared by all Instructions supported by LIEF