lief::dwarf::types

Trait DwarfType

pub trait DwarfType {
    // Provided methods
    fn name(&self) -> Result<String, Error> { ... }
    fn size(&self) -> Result<u64, Error> { ... }
    fn location(&self) -> DebugLocation { ... }
    fn is_unspecified(&self) -> bool { ... }
    fn scope(&self) -> Option<Scope<'_>> { ... }
}

Generic trait shared by all DWARF types

Provided Methods

fn name(&self) -> Result<String, Error>

Return the type’s name using either DW_AT_name or DW_AT_picture_string (if any)

fn size(&self) -> Result<u64, Error>

Return the size of the type or an error if it can’t be computed.

This size should match the equivalent of sizeof(Type).

fn location(&self) -> DebugLocation

Return the debug location where this type is defined.

fn is_unspecified(&self) -> bool

Whether this type is a DW_TAG_unspecified_type.

fn scope(&self) -> Option<Scope<'_>>

The scope in which this function is defined