Trait lief::generic::Symbol

pub trait Symbol {
    // Provided methods
    fn name(&self) -> String { ... }
    fn value(&self) -> u64 { ... }
    fn size(&self) -> u64 { ... }
}

Trait shared by all the symbols in executable formats

Provided Methods

fn name(&self) -> String

Symbol’s name

fn value(&self) -> u64

Symbol’s value whose interpretation depends on the symbol’s kind. Usually this is the address of the symbol though.

fn size(&self) -> u64

Size of the symbol (can be 0)