Module lief::macho

Module for the Mach-O format

Modules

binary

binding_info

commands

export_info

fat_binary

header

relocation

section

symbol

Structs

Binary

This is the main interface to read and write Mach-O binary attributes.

ExportInfo

This structure represents an export (info) in a Mach-O binary

FatBinary

This structure represents a FAT Mach-O

Header

Structure that represents the main Mach-O header (at the beginning of the file)

Section

Symbol

Structure that represents a Symbol in a Mach-O file.

Enums

BindingInfo

This enum exposes all the different types of binding operations that we can find in a Mach-O binary. BindingInfo::Dyld exposes the bindings info wrapped in the LC_DYLD_INFO command while BindingInfo::Chained exposes the new chained bindings implemented in the DYLD_CHAINED_FIXUPS command.

Commands

Enum that wraps all the different Mach-O load commands (LC_xxx). Note that all these commands implements the trait: Command

Relocation

Enum that represents the different to encode/represent a relocation in a Mach-O file