Module lief::macho::binding_info

Structs

Chained

This structure represents a binding operation coming from chained binding command: LC_DYLD_CHAINED_FIXUPS

Dyld

This structure represents a binding operation coming from binding bytecode of LC_DYLD_INFO

Generic

Enums

BINDING_CLASS

BIND_TYPES

BindingInfo

This enum exposes all the different types of binding operations that we can find in a Mach-O binary. BindingInfo::Dyld exposes the bindings info wrapped in the LC_DYLD_INFO command while BindingInfo::Chained exposes the new chained bindings implemented in the DYLD_CHAINED_FIXUPS command.

CHAINED_FORMAT

Traits

AsGeneric

Generic trait shared by all BindingInfo items