lief::macho

Module commands

Modules

atom_info

build_version

code_signature

code_signature_dir

data_in_code

dyld_chained_fixups

dyld_environment

dyld_export_trie

dyldinfo

dylib

dylinker

dynamic_symbol_command

encryption_info

function_variant_fixups

function_variants

functionstarts

linker_opt_hint

main_cmd

note

routine

rpath

segment

segment_split_info

source_version

sub_client

sub_framework

symbol_command

thread_command

two_level_hints

unknown

uuid

version_min

Structs

AtomInfo

Structure which represents the LC_ATOM_INFO command

BuildVersion

Structure that represents the LC_BUILD_VERSION command

CodeSignature

Structure that represents the LC_CODE_SIGNATURE command

CodeSignatureDir

Structure that represents the LC_DYLIB_CODE_SIGN_DRS command

DataInCode

Structure that represents the LC_DATA_IN_CODE command

DyldChainedFixups

Structure that represents the LC_DYLD_CHAINED_FIXUPS command

DyldEnvironment

Structure that represents a LC_DYLD_ENVIRONMENT command which is used by the Mach-O linker/loader to initialize an environment variable

DyldExportsTrie

Structure that represents the LC_DYLD_EXPORTS_TRIE command

DyldInfo

Structure that represents the LC_DYLD_INFO and LC_DYLD_INFO_ONLY commands

Dylib

Structure which represents a library dependency

Dylinker

Structure that represents the Mach-O linker, also named loader. Most of the time, Dylinker::name should return /usr/lib/dyld

DynamicSymbolCommand

Structure that represents the LC_DYSYMTAB command.

EncryptionInfo

Structure that represents the LC_ENCRYPTION_INFO / LC_ENCRYPTION_INFO_64 commands

FunctionStarts

Structure which represents the LC_FUNCTION_STARTS command

FunctionVariantFixups

Structure which represents the LC_FUNCTION_VARIANT_FIXUPS command

FunctionVariants

Class representing the LC_FUNCTION_VARIANTS load command.

Generic

LinkerOptHint

Structure which represents the LC_LINKER_OPTIMIZATION_HINT command

Main

Structure that represent the LC_MAIN command. This kind of command can be used to determine the entrypoint of an executable.

Note

Class that represent the LC_NOTE command.

RPath

Structure that represents the LC_RPATH command.

Routine

Class that represents the LC_ROUTINE/LC_ROUTINE64 commands. Accodring to the Mach-O loader.h documentation:

Segment

Class which represents a LC_SEGMENT/LC_SEGMENT_64 command

SegmentSplitInfo

Structure that represents the LC_SEGMENT_SPLIT_INFO command

SourceVersion

Class that represents the MachO LC_SOURCE_VERSION This command is used to provide the version of the sources used to build the binary.

SubClient

Class that represents the SubClient command. Accodring to the Mach-O loader.h documentation:

SubFramework

Class that represents the SubFramework command. Accodring to the Mach-O loader.h documentation:

SymbolCommand

Structure that represents the LC_SYMTAB command

ThreadCommand

Structure that represents the LC_THREAD / LC_UNIXTHREAD commands and that can be used to get the binary entrypoint when the LC_MAIN is not present

TwoLevelHints

Structure which represents the LC_TWOLEVEL_HINTS command

UUID

Structure that represents the LC_UUID command

Unknown

Generic structure when the command is not recognized by LIEF (e.g private LC_xxx command)

VersionMin

Structure that wraps the LC_VERSION_MIN_MACOSX, LC_VERSION_MIN_IPHONEOS, ... commands.

Enums

Commands

Enum that wraps all the different Mach-O load commands (LC_xxx). Note that all these commands implements the trait: Command

LoadCommandTypes

Traits

Command

Trait shared by all the load command: Commands

Type Aliases

CommandsIter

Libraries